Refactor/#22 type change

+++ src/main/java/com/petplate/petplate/petdailymeal/domain/entity/DailyMeal.java
@@ -31,7 +31,7 @@ public class DailyMeal {
     private Long id;
 
     @Column(nullable = false)
-    private float kcal;
+    private double kcal;
 
     @CreatedDate
     @Column(name = "created_at", nullable = false, updatable = false, columnDefinition = "TIMESTAMP")
